--- app/cdap/components/ConnectionList/Components/TabLabelCanBrowse/types.ts
@@ -1,36 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import { Ref } from 'react';
-
-export interface ITabLabelItemProps {
-  myLabelRef: Ref<HTMLSpanElement>;
-  label: string;
-  count: number;
-  labelTestId?: string;
-}
-
-export interface ITabLabelProps extends ITabLabelItemProps {
-  icon: JSX.Element;
-  labelContainerTestId?: string;
-}
-
-export interface ITabLabelCanBrowseProps {
-  label: string;
-  count: number;
-  columnIndex: number;
-  icon?: JSX.Element;
-}
