--- app/cdap/components/ConnectionList/Components/TabLabelCanSample/types.ts
@@ -1,35 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import { IConnectorTabType } from 'components/ConnectionList/Components/ConnectionTabs/types';
-import { Dispatch, Ref, SetStateAction } from 'react';
-
-export interface ITabLabelSampleItemProps {
-  label: string;
-  myLabelRef: Ref<HTMLSpanElement>;
-  onExplore: (entity: IConnectorTabType) => void;
-  entity: IConnectorTabType;
-  buttonTestId: string;
-  buttonElement: JSX.Element;
-}
-
-export interface ITabLabelCanSampleProps {
-  label: string;
-  entity: IConnectorTabType;
-  initialConnectionId: string;
-  toggleLoader: (value: boolean, isError?: boolean) => void;
-  setIsErrorOnNoWorkSpace: Dispatch<SetStateAction<boolean>>;
-}
